homework/module_01/pyproject.toml
[project]
name = "module-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "pandas>=2.0.3",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=17.0.0",
    "sqlalchemy>=2.0.0",
    "tqdm>=4.67.1",
    "python-dotenv>=1.0.0",
]
